% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plot.R
\name{base_plot}
\alias{base_plot}
\title{plot to compare snppit runs}
\usage{
base_plot(
  D,
  by_year = TRUE,
  type_sep_y = 0.05,
  point_size = 1,
  legend_point_size = 2,
  point_stroke = 0.3,
  line_size = 0.08
)
}
\description{
more later
@export
}
