% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coho_genotypes}
\alias{coho_genotypes}
\title{Genotypes of 1984 coho salmon over seven years}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 571392 rows and 5 columns.
}
\usage{
coho_genotypes
}
\description{
This is a tibble.  More later
}
\details{
The important part is that it must have the columns
\code{indiv}, \code{locus}, \code{gene_copy}, \code{allele_int}.
}
\keyword{datasets}
