% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coho_metadata}
\alias{coho_metadata}
\title{Sex, year, and spawner group of 1984 coho salmon over seven years}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1984 rows and 5 columns.
}
\usage{
coho_metadata
}
\description{
This is a tibble.  More later
}
\keyword{datasets}
