% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matching_samples.R
\name{find_matching_samples}
\alias{find_matching_samples}
\title{Find matching samples to a single representative}
\usage{
find_matching_samples(genotypes, return_aliases = FALSE, ...)
}
\arguments{
\item{genotypes}{A tibble like coho_genotypes that has}

\item{...}{parameters to be passed to rubias::close_matching_samples.
Intended to be used for \code{min_frac_non_miss} and \code{min_frac_matching}}
}
\description{
Given a specification of pairs of genotypes that must be from the
same individual, this identifies connected components and returns a tibble
with one column, \code{indiv}, and another column, \code{aliases} which is a list column, that
includes, for each \code{indiv}, all the other names it is known by.
}
\examples{
# There are not actually any matching samples in coho_genotypes
# but we will just create some pairs that match by cranking
# the min_frac_matching down to 80\%
find_matching_samples(coho_genotypes, min_frac_matching = 0.80)
}
