% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genos2rubias.R
\name{genos2rubias}
\alias{genos2rubias}
\title{convert long format genotypes to rubias format}
\usage{
genos2rubias(genos, allele_column = allele_int)
}
\arguments{
\item{genos}{the long format genotypes tibble with columns
\code{indiv}, \code{locus}, \code{gene_copy} and \code{allele_int}.  Missing data
in the allele_int column must be denoted by NA.}

\item{allele_column}{unquoted name of the column holding the alleles to
use.  By default this is \code{allele_int}.}

\item{missing}{quoted symbol used to denote missing data}
}
\description{
convert long format genotypes to rubias format
}
