% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_two_snppit_results.R
\name{partition_two_snppit_results}
\alias{partition_two_snppit_results}
\title{Partition the results of two SNPPIT runs by comparisons of their features}
\usage{
partition_two_snppit_results(Run1, Run2, FDR1 = 0.01, FDR2 = 0.01)
}
\arguments{
\item{Run1}{the output of \code{\link[=slurp_snppit]{slurp_snppit()}} for the first (typically with sex and date) snppit run.}

\item{Run2}{the output of \code{\link[=slurp_snppit]{slurp_snppit()}} or \code{\link[=reformat_no_sex_or_date_results]{reformat_no_sex_or_date_results()}} for the second
(typically \emph{without} sex and date) snppit run.}

\item{FDR1}{the FDR threshold to consider for Run1 (defaults to 0.01).}

\item{FDR2}{the FDR threshold to consider for Run2 (defaults to 0.01).}
}
\description{
This function is intended to break down the results of two SNPPIT runs
into non-overlapping groups of kids.  The primary purpose for this is
to compare runs with sex-and-date to runs without the sex and date.  After
doing this you can use a later function to pick out the trio from either
run1 or run2 according to which category it is in.
}
\details{
This is implemented in terms of bit-masks.  Here are the bits, given by the
"1" value.  We set this up so that good trios shared by both runs will be 0s.
\itemize{
\item 1: Parents not identified (NA) in Run1
\item 2: Parents not identified (NA) in Run2
\item 4: Max posterior trio in Run1 is not C_Se_Se
\item 8: Max posterior trio in Run2 is not C_Se_Se
\item 16: FDR > FDR1 in Run1
\item 32: FDR > FDR2 in Run2
\item 64: Run1 and Run2 parent pairs share exactly 0 individuals
\item 128: Run1 and Run2 parent pairs share exactly 1 individual
}

If neither of bits 1, 2, 64, or 128 are set, then there were two parents matching between the groups.
}
