% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_snppit.R
\name{prep_for_snppit}
\alias{prep_for_snppit}
\title{prepare long-format genotypes for snppit}
\usage{
prep_for_snppit(
  G,
  S,
  use_spawn_date = TRUE,
  use_sex = TRUE,
  minAge = 2,
  maxAge = 4,
  geno_err = 0.005,
  outf = "snippit_out.txt"
)
}
\arguments{
\item{G}{a long format data frame of SNPs.  It must have, at a minimum, the columns
\code{Indiv}, \code{Locus}, \code{gene_copy}, \code{snp_as_int}.  Missing data should be represented as
a 0.}

\item{S}{a data frame of meta data with the columns Indiv,   SEX, and    collection_date}
}
\description{
More later...
}
