% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_to_compare_snippit_runs.R
\name{prep_to_compare_snppit_runs}
\alias{prep_to_compare_snppit_runs}
\title{Prepare SNPPIT runs to compare them}
\usage{
prep_to_compare_snppit_runs(
  D,
  type_levels = c("nSD_full", "nSD_dump10", "wSD_full", "wSD_dump10"),
  GTseqed = sex_and_date$Indiv,
  GT_hi_miss = gt_miss10p
)
}
\description{
More later
}
