% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snppit_infile.R
\name{prepare_snppit_infile}
\alias{prepare_snppit_infile}
\title{prepare long-format genotypes for snppit}
\usage{
prepare_snppit_infile(
  G,
  S,
  use_spawner_group = TRUE,
  use_sex = TRUE,
  min_age = 2,
  max_age = 4,
  geno_err = 0.005,
  outf = tempfile()
)
}
\arguments{
\item{G}{a long format data frame of SNPs.  It must have, at a minimum, the columns
\code{indiv}, \code{locus}, \code{gene_copy}, \code{allele_int}.  Missing data should be represented as
NA.}

\item{S}{a data frame of meta data with the columns indiv,   sex, and    spawner_group}
}
\description{
More later...
}
\examples{
prep_for_snppit(G = coho_genotypes, S = coho_metadata)
}
