% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_snippit.R
\name{run_snppit}
\alias{run_snppit}
\title{prepare data for SNPPIT and run it}
\usage{
run_snppit(genotypes, metadata, additional_args = "", outdir = tempfile(), ...)
}
\arguments{
\item{genotypes}{A tibble of genotypes.  Should have been reoranized if there were matching samples.}

\item{metadata}{A tibble of metadata.  Should have been reoranized if there were matching samples.}

\item{additional_args}{A quoted string of additional (i.e. beyond the -f option) options and arguments
to pass to snppit.}

\item{outdir}{The directory to run SNPPIT in and in which to write the output.}

\item{...}{Additional arguments to pass to \code{prepare_snppit_infile()}.  Don't pass \code{outf} here, though!}
}
\description{
This returns the directory where it ran, so that you can slurp stuff out of there.
}
