% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurp_snppit.R
\name{slurp_snppit}
\alias{slurp_snppit}
\title{Read the output from snppit into a format useful for comparison}
\usage{
slurp_snppit(DIR, S)
}
\arguments{
\item{DIR}{the directory where the output lives}

\item{S}{the tibble of sex and date.  Must have columns
indiv,   sex, spawner_group, and hatchery}
}
\description{
More later...
}
