% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_to_compare_snippit_runs.R
\name{prep_to_compare_snppit_runs}
\alias{prep_to_compare_snppit_runs}
\title{Prepare SNPPIT runs to compare them}
\usage{
prep_to_compare_snppit_runs(L)
}
\arguments{
\item{L}{a list of outputs from different snppit runs.  These should be named by whatever
you want to call each run.}
}
\description{
More later
}
