% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_snppit_infile.R
\name{prepare_snppit_infile}
\alias{prepare_snppit_infile}
\title{prepare long-format genotypes for snppit}
\usage{
prepare_snppit_infile(
  G,
  S,
  use_spawner_group = TRUE,
  use_sex = TRUE,
  min_age = 1,
  max_age = 6,
  geno_err = 0.005,
  outf = tempfile()
)
}
\arguments{
\item{G}{a long format data frame of SNPs.  It must have, at a minimum, the columns
\code{indiv}, \code{locus}, \code{gene_copy}, \code{allele_int}.  Missing data should be represented as
NA.}

\item{S}{a data frame of meta data with the columns \code{indiv}, \code{year}, \code{sex}, \code{spawner_group},
and \code{hatchery}}
}
\value{
Function does not return a value of any importance.
The side effect of the function is to write out a snppit input file!
}
\description{
The meta data, S, determines who gets put into the snppit infile via a left
join.  So, if it turns out that we are requsting the addition of individuals
that are not in \code{G}, this will throw an error.
}
\details{
This function is called within \code{\link[=run_snippit]{run_snippit()}}.
}
\examples{
prepare_snppit_infile(G = coho_genotypes, S = coho_metadata)
}
