% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_no_sex_or_date_results.R
\name{reformat_no_sex_or_date_results}
\alias{reformat_no_sex_or_date_results}
\title{After slurping in a SNPPIT run done without sex or date constraints, harmonize the declared sex and dates with the output}
\usage{
reformat_no_sex_or_date_results(D)
}
\arguments{
\item{D}{the output of  \code{\link[=slurp_snppit]{slurp_snppit()}}.}
}
\value{
A tibble where the males have been placed in the pa columns and the females ahve been placed in the ma columns.
}
\description{
The main thing going on here is to make sure that the pa column are males and the ma column are females
and to declare a reported sex mismatch if they are both males or females.
}
