% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganize_matching_samples.R
\name{reorganize_matching_samples}
\alias{reorganize_matching_samples}
\title{Relabel matching genotypes to a single ID and reorganize their years and spawner groups}
\usage{
reorganize_matching_samples(genotypes, metadata, clusters)
}
\arguments{
\item{genotypes}{the genotypes}

\item{metadata}{the meta data}

\item{clusters}{the tibble with the matching sample clusters.}
}
\value{
A list with a variety of components:  (add later...)
}
\description{
If there are matching samples, we want to make sure that we have just a single
ID for each cluster, and that the years and spawner groups get organized
so that SNPPIT can read all that information in.
}
